[TM4C123_GPIO_LCD1602] Removing MISRAM 10.1 and 10.3 Warnings Actualizar TM4C123_GPIO_LCD1602.launch, main.c y 52 más archivos...

+++ TM4C123_GPIO_LCD1602/main.c
@@ -52,8 +52,8 @@ volatile uint32_t u32Update=0;
 volatile uint32_t u32Counter=0;
 volatile uint32_t u32Priority=0;
 
-uint8_t LCD1602_pu8Buffer1[LCD1602_COLUMN_MAX+1]=" INDEVICE TIVAC ";
-uint8_t LCD1602_pu8Buffer2[LCD1602_COLUMN_MAX+1]="  LCD1602 SW:   ";
+uint8_t LCD1602_pu8Buffer1[LCD1602_COLUMN_MAX+1u]=" INDEVICE TIVAC ";
+uint8_t LCD1602_pu8Buffer2[LCD1602_COLUMN_MAX+1u]="  LCD1602 SW:   ";
 GPIO_nBUS enBus=GPIO_enAPB;
 
 volatile uint32_t vu32Refresh=0;
@@ -74,7 +74,7 @@ int main(void)
     SYSEXC__vInit((SYSEXC_nINTERRUPT)(SYSEXC_enINVALID|SYSEXC_enDIV0|
             SYSEXC_enOVERFLOW|SYSEXC_enUNDERFLOW),SYSEXC_enPRI7);
     SYSCTL__enInit();/* system clock 80MHz*/
-    SysTick__enInitUs(10,SCB_enSHPR0);
+    SysTick__enInitUs(10.0f,SCB_enSHPR0);
     EEPROM__enInit();
     MAIN_vInitGPIO();
     LCD1602__enInit();
@@ -89,52 +89,52 @@ int main(void)
         psSW2=GPIOF_APB_GPIODATA_MASK;
     }
     LCD1602__enReloadScreenDirect();
-    u8ColumnCurrent=0;
-    u8Column=0;
-    u8Row=0;
-    while(1)
+    u8ColumnCurrent=0u;
+    u8Column=0u;
+    u8Row=0u;
+    while(1u)
     {
 
         fTimeSystickStart = SysTick__fGetTimeUs();
-        if(vu32Refresh == (enSW1Pin|enSW2Pin))
+        if(vu32Refresh == (uint32_t)((uint32_t)enSW1Pin|(uint32_t)enSW2Pin))
         {
-            u8Column=13;
-            u8Row=0;
-            LCD1602__enWriteStringBufferSection((char*)LCD1602_pu8Buffer2,"BT",&u8Column,&u8Row,&u8Counter,0,15,0,0);
+            u8Column=13u;
+            u8Row=0u;
+            LCD1602__enWriteStringBufferSection((char*)LCD1602_pu8Buffer2,"BT",&u8Column,&u8Row,&u8Counter,0u,15u,0u,0u);
         }
         else if(vu32Refresh == (enSW1Pin))
         {
-            u8Column=13;
-            u8Row=0;
-            LCD1602__enWriteStringBufferSection((char*)LCD1602_pu8Buffer2,"1 ",&u8Column,&u8Row,&u8Counter,0,15,0,0);
+            u8Column=13u;
+            u8Row=0u;
+            LCD1602__enWriteStringBufferSection((char*)LCD1602_pu8Buffer2,"1 ",&u8Column,&u8Row,&u8Counter,0u,15u,0u,0u);
         }
         else if(vu32Refresh == (enSW2Pin))
         {
-            u8Column=13;
-            u8Row=0;
-            LCD1602__enWriteStringBufferSection((char*)LCD1602_pu8Buffer2,"2 ",&u8Column,&u8Row,&u8Counter,0,15,0,0);
+            u8Column=13u;
+            u8Row=0u;
+            LCD1602__enWriteStringBufferSection((char*)LCD1602_pu8Buffer2,"2 ",&u8Column,&u8Row,&u8Counter,0u,15u,0u,0u);
         }
         else
         {
-            u8Column=13;
-            u8Row=0;
-            LCD1602__enWriteStringBufferSection((char*)LCD1602_pu8Buffer2,"NN",&u8Column,&u8Row,&u8Counter,0,15,0,0);
+            u8Column=13u;
+            u8Row=0u;
+            LCD1602__enWriteStringBufferSection((char*)LCD1602_pu8Buffer2,"NN",&u8Column,&u8Row,&u8Counter,0u,15u,0u,0u);
         }
         u8Column=u8ColumnCurrent;
-        u8Row=0;
-        LCD1602__enPrintfSection((char*)LCD1602_pu8Buffer1,&u8Column,&u8Row,(uint8_t*)&u8Counter,0,15,0,0);
+        u8Row=0u;
+        LCD1602__enPrintfSection((char*)LCD1602_pu8Buffer1,&u8Column,&u8Row,(uint8_t*)&u8Counter,0u,15u,0u,0u);
         u8Column=u8ColumnCurrent;
-        u8Row=1;
-        LCD1602__enPrintfSection((char*)LCD1602_pu8Buffer2,&u8Column,&u8Row,(uint8_t*)&u8Counter,0,15,1,1);
-        if(u8ColumnCurrent>=15)
-            u8ColumnCurrent=0;
+        u8Row=1u;
+        LCD1602__enPrintfSection((char*)LCD1602_pu8Buffer2,&u8Column,&u8Row,(uint8_t*)&u8Counter,0u,15u,1u,1u);
+        if(u8ColumnCurrent>=15u)
+            u8ColumnCurrent=0u;
         else
             u8ColumnCurrent++;
 
 
         fTimeSystickEnd = SysTick__fGetTimeUs();
         fTimeSystickEnd-= fTimeSystickStart;
-        fTimeSystickEnd=1000000.0-fTimeSystickEnd;
+        fTimeSystickEnd=1000000.0f-fTimeSystickEnd;
         SysTick__vDelayUs(fTimeSystickEnd);
 
     }
@@ -155,7 +155,7 @@ void MAIN_vInitGPIO(void)
     GPIO__enSetDigitalConfig(GPIO_enGPIOF0,GPIO_enCONFIG_INPUT_2MA_OPENDRAIN_PULLUP);
     GPIO__enSetDigitalConfig(GPIO_enGPIOF4,GPIO_enCONFIG_INPUT_2MA_OPENDRAIN_PULLUP);
 
-    GPIO__vSetData(GPIO_enPORTF,(GPIO_nPIN) (GPIO_enPIN1|GPIO_enPIN2|GPIO_enPIN3), 0);
+    GPIO__vSetData(GPIO_enPORTF,(GPIO_nPIN) (GPIO_enPIN1|GPIO_enPIN2|GPIO_enPIN3), 0u);
 
     GPIO__vClearInterrupt(GPIO_enPORTF,  (GPIO_nPIN)(GPIO_enPIN0|GPIO_enPIN4));
     GPIO__vEnInterruptConfig(GPIO_enPORTF, (GPIO_nPIN)(GPIO_enPIN0|GPIO_enPIN4), GPIO_enINT_CONFIG_EDGE_BOTH);
@@ -165,16 +165,16 @@ void MAIN_vInitGPIO(void)
 
 void MAIN_vIsrSW1(void)
 {
-    if(psSW1->DATA_MASK[enSW1Pin]==0)
+    if(psSW1->DATA_MASK[enSW1Pin]==0u)
     {
         GPIO__vSetData(GPIO_enPORTF,(GPIO_nPIN) (GPIO_enPIN1), GPIO_enPIN1);
-        vu32Refresh|=enSW1Pin;
+        vu32Refresh|=(uint32_t)enSW1Pin;
     }
     else
     {
 
-        GPIO__vSetData(GPIO_enPORTF,(GPIO_nPIN) (GPIO_enPIN1), 0);
-        vu32Refresh&=~enSW1Pin;
+        GPIO__vSetData(GPIO_enPORTF,(GPIO_nPIN) (GPIO_enPIN1), 0u);
+        vu32Refresh&=~(uint32_t)enSW1Pin;
     }
 }
 
@@ -182,15 +182,15 @@ void MAIN_vIsrSW1(void)
 void MAIN_vIsrSW2(void)
 {
 
-    if(psSW2->DATA_MASK[enSW2Pin]==0)
+    if(psSW2->DATA_MASK[(uint32_t)enSW2Pin]==0u)
     {
         GPIO__vSetData(GPIO_enPORTF,(GPIO_nPIN) (GPIO_enPIN2), GPIO_enPIN2);
-        vu32Refresh|=enSW2Pin;
+        vu32Refresh|=(uint32_t)enSW2Pin;
     }
     else
     {
 
         GPIO__vSetData(GPIO_enPORTF,(GPIO_nPIN) (GPIO_enPIN2), GPIO_enPIN0);
-        vu32Refresh&=~enSW2Pin;
+        vu32Refresh&=~(uint32_t)enSW2Pin;
     }
 }
